pyproject.toml
[tool.poetry]
name = "address-geocoder"
version = "0.1.0"
description = "Address Geocoder is a powerful Python-based solution for geocoding raw addresses"
authors = ["Mohamed Ehouran <mohamedehouran@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12"
db-dtypes = "1.3.1"
numpy = "2.2.0"
pandas = "2.2.3"
pyarrow = "18.1.0"
geopy = "2.4.1"
geopandas = "^1.0.1"
python-dotenv = "^1.0.1"
click = "^8.1.8"
pandera = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
